7.1.7 Файл с утилитами переместит в папку servises и переименует его в common.js. Переместит в него вспомогательные функции из компонента offer. Удалит ненужный проп в index.js
